Rename lib/amplitude to lib/logging
