+++ components/listings/buy/BuyListing/index.js
@@ -16,7 +16,7 @@ import {
   BUYER_LANDING_SCROLL_50,
   BUYER_LANDING_SCROLL_75,
   BUYER_LANDING_SCROLL_100
-} from 'lib/amplitude'
+} from 'lib/logging'
 import {
   UnderlinedText,
   Container,
