+++ pages/listings/buy/index.js
@@ -15,7 +15,7 @@ import {
   log,
   BUYER_LANDING_PAGE,
   BUYER_LANDING_EXPLORE_LISTINGS
-} from 'lib/amplitude'
+} from 'lib/logging'
 import NextHead from 'components/shared/NextHead'
 
 const Container = styled(View)`
